+++ src/client/cl_cin.c
@@ -1102,8 +1102,8 @@ static void readQuadInfo( byte *qData ) {
 	cinTable[currentHandle].VQ0 = cinTable[currentHandle].VQNormal;
 	cinTable[currentHandle].VQ1 = cinTable[currentHandle].VQBuffer;
 
-	cinTable[currentHandle].t[0] = ( 0 - (unsigned int)cin.linbuf ) + (unsigned int)cin.linbuf + cinTable[currentHandle].screenDelta;
-	cinTable[currentHandle].t[1] = ( 0 - ( (unsigned int)cin.linbuf + cinTable[currentHandle].screenDelta ) ) + (unsigned int)cin.linbuf;
+	cinTable[currentHandle].t[0] = ( 0 - (intptr_t)cin.linbuf ) + (intptr_t)cin.linbuf + cinTable[currentHandle].screenDelta;
+	cinTable[currentHandle].t[1] = ( 0 - ( (intptr_t)cin.linbuf + cinTable[currentHandle].screenDelta ) ) + (intptr_t)cin.linbuf;
 
 	cinTable[currentHandle].drawX = cinTable[currentHandle].CIN_WIDTH;
 	cinTable[currentHandle].drawY = cinTable[currentHandle].CIN_HEIGHT;
